// src/app/core/api/models/sort-rule.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SortRule {
  direction?: 'asc' | 'desc';

  /**
   * Sortable field name
   */
  field: 'format_version' | 'format' | 'pruefidentifikator' | 'description' | 'segmentgroup_key' | 'segment_code' | 'data_element' | 'qualifier' | 'line_ahb_status' | 'line_name' | 'bedingung' | 'direction';
}